refactor(team): extract studio endpoints into broker_studio.go
Slice 2 of broker.go decomposition. Moves the studio package
generator + workflow runner: studioPackageGeneratorFn,
studioPackageGeneratorOverride, studioPackageGenerator, all
studioGenerated*/Stub* types, decodeStudioGeneratedPackage,
buildStudioFollowUpStubExecutions, studioDefaultArtifactDefinitions,
studioNormalizeArtifactDefinitions, studioArtifactIDs,
buildStudioGeneratedArtifacts, extractStudioOfferNames,
extractStudioStringSlice, firstStudioString, handleStudioGeneratePackage,
studioPackageArtifactIDs, studioPrimaryPackageArtifact,
normalizeStudioWorkflowDefinition, studioWorkflowHints,
workflowCreateConflict, uniqueStrings, workflowRunModeLabel,
mustMarshalStudioJSON, executeStudioWorkflowStub,
recordStudioWorkflowExecution, handleStudioRunWorkflow into a new
broker_studio.go.

No behavior change — pure file relocation within package team. The
local var named `provider` in handleStudioRunWorkflow is renamed to
`prov` because the relocated file imports the `provider` package
directly. broker.go shrinks from 11424 to 9886 lines (and from 12058
in the upgrade slice). Adds three characterization tests:
TestHandleStudioGeneratePackage_RequiresAuth,
TestHandleStudioRunWorkflow_RejectsNonPostMethods,
TestNormalizeStudioWorkflowDefinition_RoundTrip. Total package
coverage holds at 62.9% baseline.
