+++ internal/team/broker_studio_test.go
@@ -157,6 +157,76 @@ func TestHandleStudioGeneratePackagePersistsAction(t *testing.T) {
 	}
 }
 
+// TestHandleStudioGeneratePackage_RequiresAuth pins that the auth gate
+// applies to /studio/generate-package — accidentally registering the
+// handler outside b.requireAuth would let any caller invoke an LLM
+// generation behind the broker.
+func TestHandleStudioGeneratePackage_RequiresAuth(t *testing.T) {
+	b := newTestBroker(t)
+	srv := httptest.NewServer(b.requireAuth(b.handleStudioGeneratePackage))
+	defer srv.Close()
+
+	resp, err := http.Post(srv.URL, "application/json", strings.NewReader("{}"))
+	if err != nil {
+		t.Fatalf("request: %v", err)
+	}
+	defer resp.Body.Close()
+	if resp.StatusCode != http.StatusUnauthorized {
+		t.Errorf("expected 401, got %d", resp.StatusCode)
+	}
+}
+
+// TestHandleStudioRunWorkflow_RejectsNonPostMethods pins the method gate.
+// The endpoint mutates broker state (records actions, runs workflows) so
+// drifting to allow GET/PUT/DELETE could expose the operation to CSRF or
+// to GETs from log-tailing tools. Lock the contract here.
+func TestHandleStudioRunWorkflow_RejectsNonPostMethods(t *testing.T) {
+	b := newTestBroker(t)
+	srv := httptest.NewServer(http.HandlerFunc(b.handleStudioRunWorkflow))
+	defer srv.Close()
+
+	for _, method := range []string{http.MethodGet, http.MethodPut, http.MethodDelete, http.MethodPatch} {
+		req, _ := http.NewRequest(method, srv.URL, nil)
+		resp, err := http.DefaultClient.Do(req)
+		if err != nil {
+			t.Fatalf("%s: request: %v", method, err)
+		}
+		_ = resp.Body.Close()
+		if resp.StatusCode != http.StatusMethodNotAllowed {
+			t.Errorf("%s: expected 405, got %d", method, resp.StatusCode)
+		}
+	}
+}
+
+// TestNormalizeStudioWorkflowDefinition_RoundTrip pins the three branches
+// of the normaliser: empty / null → nil, JSON-string-wrapped → unwrapped,
+// raw JSON object → passed through.
+func TestNormalizeStudioWorkflowDefinition_RoundTrip(t *testing.T) {
+	cases := []struct {
+		name string
+		in   string
+		want string
+	}{
+		{"empty", "", ""},
+		{"whitespace only", "   ", ""},
+		{"literal null", "null", ""},
+		{"json-string wrap", `"{\"steps\":[]}"`, `{"steps":[]}`},
+		{"raw object", `{"steps":[]}`, `{"steps":[]}`},
+		{"trim outer", "  {\"a\":1}  ", `{"a":1}`},
+	}
+	for _, tc := range cases {
+		t.Run(tc.name, func(t *testing.T) {
+			got, err := normalizeStudioWorkflowDefinition(json.RawMessage(tc.in))
+			if err != nil {
+				t.Fatalf("unexpected error: %v", err)
+			}
+			if string(got) != tc.want {
+				t.Errorf("want %q, got %q", tc.want, string(got))
+			}
+		})
+	}
+}
+
 func TestHandleMemoryRoundTripScopedStudioRecords(t *testing.T) {
 	b := newTestBroker(t)
